requirements.txt
anytree==2.6.0
attrs==19.3.0
backtrader>=1.9.70.122
cryptography==2.1.4
numpy==1.16.2
PyYAML==3.12
scikit-learn==0.22
scipy==1.4.1
SQLAlchemy==1.3.12
tensorboard>=2.0.2
tensorflow>=2.0.0
tensorflow-datasets==1.3.0
tensorflow-estimator>=2.0.1
tensorflow-gpu>=2.0.0
tensorflow-metadata==0.15.1
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.39.0
tushare==1.2.48
urllib3==1.22
